+++ refinery-core/src/main/java/org/wikimedia/analytics/refinery/core/maxmind/DatabaseReaderCache.java
@@ -17,32 +17,40 @@
 package org.wikimedia.analytics.refinery.core.maxmind;
 
 import java.io.IOException;
-import java.util.concurrent.ExecutionException;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 import com.maxmind.db.NodeCache;
+import com.github.benmanes.caffeine.cache.Caffeine;
+import com.github.benmanes.caffeine.cache.Cache;
 
 public class DatabaseReaderCache implements NodeCache {
     private Cache<Integer, JsonNode> cache;
 
     /**
-     * This initializer adds a Guava cache to the MaxMind DB reader.
-     * This cache is limited in size. Least recently used items are removed first.
-     * [[https://github.com/google/guava/wiki/CachesExplained]]
+     * This initializer adds a Caffeine cache to the MaxMind DB reader.
+     * This cache is limited in size.
+     * [[https://github.com/ben-manes/caffeine]]
      *
      * @param cacheSize The maximum number of item in the cache
      */
     public DatabaseReaderCache(int cacheSize) {
-        cache = CacheBuilder.newBuilder().maximumSize(cacheSize).build();
+        cache = Caffeine.newBuilder().maximumSize(cacheSize).build();
     }
 
-    public JsonNode get(int key, Loader loader) throws IOException {
-        try {
-            return cache.get(key, () -> loader.load(key));
-        } catch (ExecutionException e) {
-            throw new IOException("Failure when loading from Maxmind DatabaseReaderCache", e);
+    public final JsonNode get(int key, Loader loader) throws IOException {
+        // NOTE: We use the getIfPresent/put strategy instead of the
+        // cache.get(key, loadingMethod) one as the latter fails on our unit-tests,
+        // leading to a never-ending loop when managing the cache.
+        //
+        // The reason is Maxmind can do recursive calls to the cache which is explicitly not supported by Caffeine.
+        // https://github.com/maxmind/MaxMind-DB-Reader-java/blob/main/src/main/java/com/maxmind/db/Decoder.java#L116
+        // https://github.com/ben-manes/caffeine/wiki/Faq#recursive-computations
+        //
+        JsonNode result = cache.getIfPresent(key);
+        if (result == null) {
+            result = loader.load(key);
+            cache.put(key, result);
         }
+        return result;
     }
 }